[WIP] Add vmware box
This adds a new make target boot2docker_vmware.box that builds a
box for vmware_desktop providers. It does not yet work properly, because
on first boot for some reason sda1 is not mounted automatically and
because vagrant fails to configure networking.
It also requires an ugly hack to boot from the iso:
Because packers vmware-iso builder has no option to keep the iso file
it used during build and it overrides the ide1:0 devise, we need to add
a second ide1:1 cdrom that points to the proper iso.